refactor: feed 펙토리 디렉터리 변경
