+++ src/main/java/one/colla/feed/common/application/FeedService.java
@@ -20,8 +20,8 @@ import one.colla.feed.common.application.dto.response.ReadFeedDetails;
 import one.colla.feed.common.domain.Feed;
 import one.colla.feed.common.domain.FeedRepository;
 import one.colla.feed.common.domain.FeedType;
-import one.colla.feed.common.factory.ReadFeedDetailsFactory;
-import one.colla.feed.common.factory.ReadFeedDetailsFactoryProvider;
+import one.colla.feed.factory.ReadFeedDetailsFactory;
+import one.colla.feed.factory.ReadFeedDetailsFactoryProvider;
 import one.colla.file.domain.AttachmentType;
 import one.colla.global.exception.CommonException;
 import one.colla.global.exception.ExceptionCode;
